+++ ts_deepscan/analyser/Dataset.py
@@ -1,9 +1,9 @@
 # SPDX-FileCopyrightText: 2020 EACG GmbH
 #
 # SPDX-License-Identifier: Apache-2.0
-
 import json
 import shutil
+import typing as t
 
 from pathlib import Path
 
@@ -15,14 +15,20 @@ class Dataset(object):
         self.__data = None
 
     @property
+    def _licenses(self) -> t.Iterable[t.Tuple[str, dict]]:
+        lics = Path(__file__).parent / 'licenses.json'
+        with lics.open('r') as fp:
+            licenses = json.load(fp)
+            return licenses.items()
+
+    @property
     def data(self):
         if not self.__data:
             raise Exception('Dataset is not loaded.')
 
         return self.__data
 
-
-    def preload(self, rebuildcache=False):
+    def load(self, rebuildcache=False):
         path = self.__path
         datasetpath = self.__datasetpath
 
@@ -47,19 +53,6 @@ class Dataset(object):
                 self.__data = json.load(fp)
 
 
-    def load(self, rebuildcache=False):
-        from .textutils import create_doc
-
-        if not self.__data:
-            self.preload(rebuildcache)
-
-        for _, val in self.__data.items():
-            docpath = val['path']
-            doc = create_doc()
-            doc.from_disk(docpath)
-            val['doc'] = doc
-
-
     def _build(self):
         data = self._build_data()
         with open(self.__datasetpath, 'w') as fp:
@@ -70,29 +63,27 @@ class Dataset(object):
 
     def _build_data(self):
         data = {}
-        lics = Path(__file__).parent / 'licenses.json'
 
-        with lics.open('r') as fp:
-            licenses = json.load(fp)
+        for key, lic in self._licenses:
+            text = lic.get('text', None)
+            if text:
+                data[key] = _build_entry(text)
+                data[key]['name'] = lic.get('name', '')
+                data[key]['aliases'] = lic.get('aliases', [])
 
-            for key, lic in licenses.items():
-                text = lic.get('text', None)
-                if text:
-                    data[key] = self._build_entry(key, text)
-                    data[key]['name'] = lic.get('name', '')
-                    data[key]['aliases'] = lic.get('aliases', [])
-
-            return data
+        return data
 
 
-    def _build_entry(self, key, text):
-        from .textutils import create_doc, compute_hash
-        docpath = self.__path / '{}.bin'.format(key)
 
-        doc = create_doc(text)
-        doc.to_disk(docpath)
+def _build_entry(text):
+    from .textutils import create_doc, compute_hash, similarity_tokens_from_doc
+#       docpath = self.__path / '{}.bin'.format(key)
 
-        return {
-            'path': str(docpath),
-            'hash': compute_hash(doc)
-        }
+    doc = create_doc(text)
+#        doc.to_disk(docpath)
+
+    return {
+        #'path': str(docpath),
+        'hash': compute_hash(doc),
+        'orths': [t.orth for t in similarity_tokens_from_doc(doc)]
+    }